Move multi-layer analysis to default
This is work towards #802

The multi-layer image analysis can be applied to any container
image, not just docker images. Hence we move that piece of code
to the default folder. This commit will break operation, so it
must be applied with the rest of the changeset.
